requirements.txt
flask
httpx
mysql-connector-python
psycopg2-binary
python-dotenv
werkzeug